feat(2fa email): update user data after 2fa change
